refactor previous code
