+++ decorators/base_decorator.rb
@@ -1,4 +1,4 @@
-require_relative './nameable'
+require_relative '../nameable'
 
 class BaseDecorator < Nameable
   attr_accessor :nameable
